requirements.txt
biopython==1.81
bokeh==3.3.1
contourpy==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tornado==6.4
tzdata==2023.3
xyzservices==2023.10.1
